refactor: move atoms in web
