+++ packages/jotai/src/atoms/index.ts
@@ -1,2 +1,1 @@
-export * from "./reader";
 export * from "./kiwi";
